pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Mihail Pugachev <xd393727@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.1.1"
alembic = "^1.13.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.30"}
faker = "^25.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
